+++ VKmfSoft_EHealth_API/Models/Domain/Patient/PatientScan.cs
@@ -7,7 +7,7 @@ namespace VKmfSoft_EHealth_API.Models.Domain.Patient
         [Key]
         public int Id { get; set; }
         public int PatientMedicalHistoryId { get; set; }//FK to PatientMedicalHistory
-        public required int ScanTypeId { get; set; }
+        public required byte ScanTypeId { get; set; }
         public DateTime ScanDate { get; set; }
         public int PerformingMedicalWorkerId { get; set; }
         public string? Notes { get; set; }
